Update project structure and refactoring
