+++ app/core/config.py
@@ -4,8 +4,7 @@ from pydantic_settings import BaseSettings
 class Settings(BaseSettings):
     APP_NAME: str = "SustAIn LLM"
     VERSION: str = "1.0.0"
-    API_PREFIX: str = "/api/main"
-    # Will add more setting (e.g., DATABASE_URL)
+    API_PREFIX: str = "/api/llm"
 
 
 settings = Settings()
