+++ app/model/request/prompt_request.py
@@ -1,0 +1,5 @@
+from pydantic import BaseModel
+
+# For request structure
+class PromptRequest(BaseModel):
+    input_text: str
